Tweak scopes and add new commands
