+++ app/controllers/slack/slack_installations_controller.rb
@@ -1,7 +1,7 @@
 require "securerandom"
 
 module Slack
-  class SlackInstallationsController < ApplicationController
+  class SlackInstallationsController < BaseController
     allow_unauthenticated_access only: %i[ create ]
 
     def create
@@ -41,7 +41,7 @@ module Slack
 
         SlackInstallation.create!(
           id: SecureRandom.uuid,
-          user: find_or_create_user(user_id, token),
+          user: find_or_create_slack_user(user_id, token),
           token:,
           oauth_version: :v2,
           oauth_scope:,
@@ -53,27 +53,6 @@ module Slack
         )
 
         redirect_to root_path, notice: "Slack Team created!"
-      end
-    end
-
-    private
-
-    def find_or_create_user(slack_user_id, token)
-      client = Slack::Web::Client.new(token: token)
-      user_info = client.users_info(user: slack_user_id).user.profile
-      slack_user = SlackUser.find_by(slack_user_id: slack_user_id)
-
-      if slack_user
-        slack_user.user
-      else
-        user = User.create_with(
-                      id: SecureRandom.uuid,
-                      name: user_info.real_name,
-                      email_address: user_info.email,
-                      password: SecureRandom.uuid)
-                   .find_or_create_by!(email_address: user_info.email)
-        SlackUser.create!(id: SecureRandom.uuid, slack_user_id:, user_id: user.id)
-        user
       end
     end
   end
