+++ slack-app-manifest.json
@@ -13,13 +13,17 @@
     "oauth_config": {
         "scopes": {
             "bot": [
+                "app_mentions:read",
                 "chat:write",
                 "chat:write.public",
                 "commands",
                 "team:read",
                 "users:read",
                 "users:read.email",
-                "app_mentions:read"
+                "channels:read",
+                "groups:read",
+                "im:read",
+                "mpim:read"
             ]
         }
     },
